==> analysis/scRNA_seq/01_processing/notebooks/nb_3_bone_atlas_add_demultiplexed_barcodes_and_doublet_scores.ipynb <==
import scanpy as sc 
import pandas as pd
pd.set_option('display.max_rows', 70)
pd.set_option('display.max_columns', 70)  # show all columns
pd.set_option('display.width', None)        # don’t wrap columns

adata= sc.read_h5ad('/lustre/scratch124/cellgen/haniffa/users/sm54/data/Bone_Atlas/bone_atlas_filtered_with_metadata.h5ad')

adata

# list exactly the obs columns you want to remove
cols_to_remove = [
    'Age_pf', 'Anatomical_region_pf', 'BCR_library_id',
    'Comment[library construction]', 'Comment[technical replicate group]',
    'Donor_pf', 'End_Bias', 'Factor Value[facs sorting]', 'File_pf',
    'Method_pf', 'N_cells', 'N_cells_pf', 'Organ_Orig', 'Organ_pf',
    'Processing_method', 'Processing_method_pf', 'Sample', 'Sample_ID_Copy',
    'Sample_pf', 'Sampling_Site', 'Sex', 'Sex_pf', 'Sort_id_pf',
    'Termination_method_pf', 'Unnamed: 14', 'abTCR_library_id_pf',
    'gdTCR_library_id_pf', 'Organ.1',
]

# drop them all at once from adata.obs
adata.obs.drop(columns=cols_to_remove, inplace=True)

# sanity check: ensure they’re gone
print([c for c in cols_to_remove if c in adata.obs.columns])
# should print an empty list


adata

doublets= pd.read_csv('/lustre/scratch126/cellgen/haniffa/sm54/BoneAtlasProject/data/combined_scrublet_scores_raw_bone_atlas.csv')

doublets.dtypes

# 2) Prepare doublets_df: set its index to Cell_ID so the join lines up
doublets_indexed = doublets.set_index('Cell_ID')

# 3) Join all columns from doublets_df into adata.obs, preserving all cells
adata.obs = adata.obs.join(doublets_indexed, how='left')

# 4) (Optional) Verify that new columns appeared
print([c for c in doublets.columns if c != 'Cell_ID'])

# List of columns to rename
cols_to_rename = [
    'scrublet_score',
    'predicted_doublets',
    'scrublet_cluster_score',
    'bh_pval',
    'predicted_doublets_pval'
]

# Build the rename mapping
rename_dict = {col: f"{col}_unfiltered" for col in cols_to_rename}

# Rename in place in adata.obs
adata.obs.rename(columns=rename_dict, inplace=True)


(adata.obs.head())

adata.obs['predicted_doublets_pval_unfiltered']=adata.obs['predicted_doublets_pval_unfiltered'].astype(str)
adata.obs['predicted_doublets_unfiltered']=adata.obs['predicted_doublets_unfiltered'].astype(str)

adata.obs['predicted_doublets_pval_unfiltered'].value_counts()

# add demultiplexed barcodes

import pandas as pd

dmx_barcodes= pd.read_csv('/nfs/team298/sm54/BoneAtlasProject/metadata/supercell_results/skull_barcodes.csv')

dmx_barcodes= pd.read_csv('/nfs/team298/sm54/BoneAtlasProject/metadata/supercell_results/')

dmx_barcodes


dmx_barcodes = dmx_barcodes.astype(str)


print(dmx_barcodes.dtypes)


adata.obs["Donor_Final"]= adata.obs["Donor_ID"].astype(str).copy()
adata.obs["Age_Final"]= adata.obs["Age"].astype(str).copy()

# 1) Ensure your adata.obs index is Cell_ID
#    If not, first:
# adata.obs.set_index('Cell_ID', inplace=True)

# 2) Make sure dmx_barcodes is indexed by Cell_ID as well
dmx = dmx_barcodes.set_index('Cell_ID')

# 3) Find the intersection of indices
common = adata.obs.index.intersection(dmx.index)

# 4) Assign Age_final into Donor_final for those cells
adata.obs.loc[common, 'Donor_Final'] = dmx.loc[common, 'Donor_final']
adata.obs.loc[common, 'Age_Final'] = dmx.loc[common, 'Age_final']

# 5) (Optional) Check a few
print(adata.obs.loc[common[:5], ['Donor_Final']])


adata.write_h5ad('/lustre/scratch124/cellgen/haniffa/users/sm54/data/Bone_Atlas/bone_atlas_filtered_with_metadata_scvi.h5ad')

#adata= sc.read_h5ad('/lustre/scratch126/cellgen/haniffa/sm54/BoneAtlasProject/data/bone_atlas_filtered_with_metadata_scvi.h5ad')

#dmx_barcodes= pd.read_csv('/nfs/team298/sm54/BoneAtlasProject/metadata/supercell_results/skull_barcodes.csv')

# Build lookup Series
status_map     = dmx_barcodes.set_index('Cell_ID')['status']
assignment_map = dmx_barcodes.set_index('Cell_ID')['assignment']
brc_map = dmx_barcodes.set_index('Cell_ID')['brc_code']
pcw_map = dmx_barcodes.set_index('Cell_ID')['pcw_published']
age_map = dmx_barcodes.set_index('Cell_ID')['Age_final']
donor_map = dmx_barcodes.set_index('Cell_ID')['Donor_final']

# Map into obs and fill missing with "singlet"
adata.obs['dmx_status']     = (
    adata.obs['Cell_ID']
        .map(status_map)
        .fillna('NA')
)
adata.obs['dmx_assignment'] = (
    adata.obs['Cell_ID']
        .map(assignment_map)
        .fillna('NA')
)

adata.obs['dmx_brc_code'] = (
    adata.obs['Cell_ID']
        .map(brc_map)
        .fillna('NA')
)

adata.obs['dmx_pcw_published'] = (
    adata.obs['Cell_ID']
        .map(pcw_map)
        .fillna('NA')
)

adata.obs['dmx_age_final'] = (
    adata.obs['Cell_ID']
        .map(age_map)
        .fillna('NA')
)

adata.obs['dmx_donor_final'] = (
    adata.obs['Cell_ID']
        .map(donor_map)
        .fillna('NA')
)

adata.obs['dmx_status']=adata.obs['dmx_status'].astype(str)
adata.obs['dmx_assignment']=adata.obs['dmx_assignment'].astype(str)
adata.obs['dmx_brc_code']=adata.obs['dmx_brc_code'].astype(str)
adata.obs['dmx_pcw_published'] =adata.obs['dmx_pcw_published'].astype(str)
adata.obs['dmx_age_final']=adata.obs['dmx_age_final'].astype(str)
adata.obs['dmx_donor_final']=adata.obs['dmx_age_final'].astype(str)

adata.obs['dmx_status'].value_counts()

adata.obs['dmx_donor_final'].value_counts()

list(adata.obs['Donor_Final'].unique())

list(dmx_barcodes['Donor_final'].unique())

# 0) Grab your two Series
s1 = dmx_barcodes['Donor_final'].astype(str)
s2 = adata.obs['Donor_Final'].astype(str)

# 1) Define a cleaning function
def clean_label(s: pd.Series) -> pd.Series:
    # a) Decode mojibake: interpret as latin1 → utf-8
    s = s.str.encode('latin1', errors='ignore') \
         .str.decode('utf-8', errors='ignore')
    # b) Drop trailing .0 on numeric labels
    s = s.str.replace(r'\.0$', '', regex=True)
    # # c) Replace any run of non‐alphanumeric with a single hyphen
    # s = s.str.replace(r'[^A-Za-z0-9]+', '-', regex=True)
    # # d) Strip hyphens off the ends
    s = s.str.strip('-')
    return s

# 2) Clean both
dmx_barcodes['Donor_clean'] = clean_label(s1)
adata.obs['Donor_clean']    = clean_label(s2)

# 3) See how many unique labels you now have
print("DMX uniques:", sorted(dmx_barcodes['Donor_clean'].unique()))
print("AnnData uniques:", sorted(adata.obs['Donor_clean'].unique()))

# 4) Find labels in DMX not in AnnData (and vice versa)
dmx_only = set(dmx_barcodes['Donor_clean']) - set(adata.obs['Donor_clean'])
adata_only = set(adata.obs['Donor_clean']) - set(dmx_barcodes['Donor_clean'])
print("In DMX not in AnnData:", dmx_only)
print("In AnnData not in DMX:", adata_only)


adata.obs['Donor_clean'].value_counts() 

#adata.obs['Donor_ID'] = adata.obs['Donor_ID'].astype(str).str.replace(r'\.0$', '', regex=True)


to_remove= ['2356/2372','2370/2369','2392/2394','multiplexed','2372/16403']

adata=adata[~adata.obs['Donor_clean'].isin(to_remove)]

adata.obs["Donor_clean"].value_counts()

adata.write_h5ad('/lustre/scratch124/cellgen/haniffa/users/sm54/data/Bone_Atlas/bone_atlas_filtered_with_metadata_scvi.h5ad')

adata.obs.to_csv('/nfs/team298/sm54/BoneAtlasProject/metadata/cell_metadata/bone_atlas_cell_metadata.csv')

import pandas as pd 

doubs= pd.read_csv('/lustre/scratch124/cellgen/haniffa/users/sm54/data/Bone_Atlas/combined_scrublet_scores_filtered_bone_atlas.csv')

doubs['predicted_doublets'].value_counts()

doubs['predicted_doublets_pval'].value_counts()

32255 + 1546870
