==> analysis/scRNA_seq/01_processing/notebooks/nb_5_scauto_qc.ipynb <==
import scanpy as sc 

adata= sc.read_h5ad('/lustre/scratch127/cellgen/cellgeni/tickets/tic-3721/work/scAutoQC/scautoqc-results-hemato_v1/scautoqc_pooled.h5ad')

adata.obs['pass_default'].value_counts()

adata.obs['suo22_pred'].value_counts()

adata

adata.obs['pass_auto_filter'].value_counts()

adata.obs.to

adata.obs.to_csv("/nfs/team298/sm54/BoneAtlasProject/metadata/cell_metadata/adata_scautoqc_obs.csv", index=True)
